+++ backend-rust-derive-lib/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "tucant_derive_lib"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+actix-web = { version = "4" }
+chrono = { version = "0.4", features = ["serde"] }
