+++ backend-rust-derive/Cargo.toml
@@ -13,3 +13,4 @@ syn = { version = "1", features = ["full"] }
 quote = "1"
 proc-macro2 = "1"
 prettyplease = "0.1"
+tucant_derive_lib = { path = "../backend-rust-derive-lib" }
