+++ backend-rust-derive/tests/test_1.rs
@@ -1,0 +1,8 @@
+use tucant_derive::Typescriptable;
+
+#[derive(Typescriptable)]
+struct Test { inner: i32 }
+
+
+#[derive(Typescriptable)]
+struct TestT<T> { inner: T }
